% File LPJ/man/baccini.Rd
\name{baccini}
\alias{baccini}
\title{Dataset on tropical above ground biomass (AGB) based on Baccini
  et al. (2012).}
\description{
  Half degree data frame of above ground biomass carbon from Baccini et
  al. (2012) in kgC/m^2.
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  Baccini, A., Goetz, S. J., Walker, W. S., Laporte, N. T., Sun, M., Sulla-Menashe, D., Hackler, J., Beck, P. S. A., Dubayah, R., Friedl, M. A., Samanta, S. and Houghton, R. A.: Estimated carbon dioxide emissions from tropical deforestation improved by carbon-density maps, Nature Clim. Change, 2(3), 182-185, doi:10.1038/nclimate1354, 2012.
}
