% File LPJ/man/biomes.legend.Rd
\name{biomes.legend}
\alias{biomes.legend}
\title{Legend and color code for biome classification}
\description{
  Data frame, which contains the available biomes created with
  \code{\link{biome_classification}} and color code equivalent to Smith
  et al. (2014). Each time \code{\link{biome_classification}} is called
  this data frame is updated.
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
Smith, B., Warlind, D., Arneth, A., Hickler, T., Leadley, P., Siltberg, J. and Zaehle, S.: Implications of incorporating N cycling and N limitations on primary production in an individual-based dynamic vegetation model, Biogeosciences, 11(7), 2027-2054, doi:10.5194/bg-11-2027-2014, 2014.
}
