% File LPJ/man/doy2mmdd.Rd
\name{doy2mmdd}
\alias{doy2mmdd}
\title{Day of year to date conversion}
\description{
  Returns the date as string in the format MMDD for a given day of the year
}
\usage{
doy2mmdd(doy, leap)
}
\arguments{
  \item{doy}{day of the year}
  \item{leap}{If the year is a leap year (default: FALSE)}
}
\value{
  string in the format MMDD
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
