% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exitMPI.R
\name{exitMPI}
\alias{exitMPI}
\title{The function to close MPI connection on cluster.}
\usage{
exitMPI()
}
\description{
This function will close slaves and finalize mpi.
}
\details{
The exitMPI should be use when working on MPI clusters. It must be
called at the end of you script. Be aware that when mpi is exited, it will be
no longer possible to work on MPI clusters without relaunching R.
The function is a wrapper of mpi.finalize from the Rmpi package. Check the
package manual for futher advise on using mpi.finalize and mpi.quit.
}
\examples{
\dontrun{
##### MPI CLUSTER ######
# Create a parallel setup for MPI cluster
mySetup <- setupLPJParallel(numCores= 20, clusterType = "MPI", mainDir=mainDir)
mySetup
# 20 different parameters to test (the number must be larger than the number of cores)
  # as a matrix
parameterList <- as.matrix(seq(1,5, len = 20))
colnames(parameterList) <- "common_emax"
  # as a list
dummy <- list (common_emax = NULL)
par <- seq(1,5, len = 20)
parameterList <- vector("list", length(par))
for (i in 1:length(par)) {
  dummy$common_emax <- par[i]
  parameterList[[i]] <- dummy
}

# Run the model
result <-   runLPJ(mySetup,  settings= settings, parameterList = parameterList )

# At the end of the script call exitMPI()
exitMPI()
}
}
\seealso{
\url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig
}
\keyword{rLPJGUESS}
