% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJDataClass.R, R/LPJSetupClass.R
\docType{methods}
\name{[,LPJData,character-method}
\alias{[,LPJData,character-method}
\alias{[,LPJSetup,character-method}
\title{Extract parts of LPJData}
\usage{
\S4method{[}{LPJData,character}(x, i, j, ..., drop = TRUE)

\S4method{[}{LPJSetup,character}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{an \code{LPJData} or an \code{LPJSetup} object}

\item{i}{a character string indicating the slot name of an LPJData class object,
an LPJ-GUESS output or a character string indicating the slot name of an LPJSetup
 class object}

\item{j}{a character string indicating a sublevel of any slot (only needed if i is provided)}

\item{...}{additional arguments (none implemented)}

\item{drop}{unused}
}
\description{
Extract parts of LPJData

Extract parts of LPJSetup
}
