% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesign.R
\name{getDesign}
\alias{getDesign}
\title{A get design function}
\usage{
getDesign(scale, list = F)
}
\arguments{
\item{scale}{a character string indicating whether the parameters are for the global or europe template}

\item{list}{a character boolean to specify the returned format. If TRUE the parameter will be returned
as a list, otherwise as a matrix.}
}
\value{
a list or a matrix with design parameter names and their values
}
\description{
This function returns the LPJ-GUESS default design for running the model
}
\examples{
design <- getDesign("global")
str(design)

design
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara
}

