% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameterList.R
\name{getParameterList}
\alias{getParameterList}
\title{A get parameter list function}
\usage{
getParameterList(scale, list = TRUE)
}
\arguments{
\item{scale}{a character string indicating whether the parameters are for the global or europe template}

\item{list}{a character boolean to specify the returned format. If TRUE the parameter will be returned
as a list, otherwise as a matrix.}
}
\value{
a list or a matrix with parameter names and their values
}
\description{
This function returns the LPJ-GUESS parameter list required for the writeTemplate function.
It also allows users to see the default values of the templates.
}
\examples{
parameterList <- getParameterList("global", list = T)
head(parameterList)
tail(parameterList)


parameterList <- getParameterList("global", list = F)
head(parameterList)
tail(parameterList)
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara
}

