% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTypeList.R
\name{getTypeList}
\alias{getTypeList}
\title{A get type list function}
\usage{
getTypeList()
}
\value{
a character vector with the default output types
}
\description{
This function returns the LPJ-GUESS output type list
}
\examples{
\dontrun{
typelist <- getTypeList()
}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
