% File LPJ/man/hyde.Rd
\name{hyde}
\alias{hyde}
\title{Fractional crop and pasture coverage for the year 2000.}
\description{
  Half degree data frame of fractional crop and pasture coverage in
  percent for the year 2000 based on Goldewijk et al. (2001, 2011).
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  Goldewijk, K. K.: Estimating global land use change over the past 300 years: The HYDE Database, Global Biogeochem. Cycles, 15(2), 417-433, doi:10.1029/1999GB001232, 2001.
Goldewijk, K. K., Beusen, A., van Drecht, G. and de Vos, M.: The HYDE 3.1 spatially explicit database of human-induced global land-use change over the past 12,000 years, Global Ecology and Biogeography, 20(1), 73-86, doi:10.1111/j.1466-8238.2010.00587.x, 2011.
}
