% File LPJ/man/isleapyear.Rd
\name{is.leapyear}
\alias{is.leapyear}
\title{Is a given year a leap year}
\description{
  Checks whether a given year is a leap year
}
\usage{
is.leapyear(year)
}
\arguments{
  \item{year}{year to check}
}
\value{
  boolean
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
