% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJDataClass.R
\docType{class}
\name{LPJData-class}
\alias{LPJData-class}
\title{A LPJData class object}
\value{
a S4 class object with two slots \itemize{
\item  runInfo contains a list with the information used to run the model such as templates and parameters
\item  dataTypes contains a list with the model outputs
}
}
\description{
This is a S4 class object to store both the LPJ-GUESS outputs
and the information that was passed to LPJ-GUESS.
}
\author{
Ramiro Silveyra Gonzalez
}
\keyword{rLPJGUESS}
