% File LPJ/man/mmdd2doy.Rd
\name{mmdd2doy}
\alias{mmdd2doy}
\title{Date conversion to day of year}
\description{
  Returns the day of a year for a given date in the format 
}
\usage{
mmdd2doy(mmdd, leap)
}
\arguments{
  \item{mmdd}{date if the form MMDD (string or numeric type)}
  \item{leap}{If the year is a leap year (default: FALSE)}
}
\value{
  day of the year
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
