% File LPJ/man/hyde.Rd
\name{mod17a3}
\alias{mod17a3}
\title{Mean GPP in kg m^-2 (2000-2014).}
\description{
  Bilinear interpolation of 30 arcsec to half degree data frame of
  gross primary production derived from satellite data (Zhao et al.,
  2005, 2006; Zhao & Running, 2010).
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  http://www.ntsg.umt.edu/project/mod17

  ftp://ftp.ntsg.umt.edu/pub/MODIS/NTSG_Products/MOD17/GeoTIFF/MOD17A3/GeoTIFF_30arcsec/

  Zhao, M., F. A. Heinsch, R. R. Nemani, and S. W. Running. (2005). Improvements
  of the MODIS terrestrial gross and net primary production global data set.
  Remote Sensing of Environment, 95: 164.176.

  Zhao, M., S. W. Running, and R. R. Nemani. (2006). Sensitivity of Moderate
  Resolution Imaging Spectroradiometer (MODIS) terrestrial primary production to
  the accuracy of meteorological reanalyses. Journal of Geophysical Research,
  111, G01002.

  Zhao, M., S. W. Running. (2010). Drought-induced reduction in global
  terrestrial net primary production from 2000 through 2009. Science, 329:
  940-943.
}
