% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPJDataClass.R
\name{[<-,LPJData,character-method}
\alias{[<-,LPJData,character-method}
\title{Replace parts of LPJData}
\usage{
\S4method{[}{LPJData,character}(x, i, j, ...) <- value
}
\arguments{
\item{x}{an \code{LPJData} object}

\item{i}{a character string indicating the LPJ-GUESS output}

\item{j}{a character string indicating a sublevel of the dataType slot (only needed if i is provided)}

\item{...}{additional arguments (none implemented)}

\item{value}{any value to create or replace and existing}
}
\description{
Replace parts of LPJData
}
