% File LPJ/man/rfs.Rd
\name{rfs}
\alias{rfs}
\title{Rainfall seasonlality, based on CRU (1977-2006) and PGF (1977-2006).}
\description{
Half degree resolution data frame containing rainfall seasonality based
on two climate datasets: CRU and Justin Sheffields
"Princeto Global Forcing" (PGF).
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{

}
