% File LPJ/man/saatchi.Rd
\name{saatchi}
\alias{saatchi}
\title{Dataset on tropical total biomass C based on Saatchi
  et al. (2011).}
\description{
  Half degree data frame of total biomass carbon from Saatchi et
  al. (2011) in kgC/m^2.
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  Saatchi, S. S., Harris, N. L., Brown, S., Lefsky, M., Mitchard, E. T. A., Salas, W., Zutta, B. R., Buermann, W., Lewis, S. L., Hagen, S., Petrova, S., White, L., Silman, M. and Morel, A.: Benchmark map of forest carbon stocks in tropical regions across three continents, PNAS, 108(24), 9899-9904, doi:10.1073/pnas.1019576108, 2011.
}
