% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupLPJParallel.R
\name{setupLPJParallel}
\alias{setupLPJParallel}
\title{The function to create a setup for parallel runs of the LPJ-GUESS}
\usage{
setupLPJParallel(numCores, clusterType, mainDir)
}
\arguments{
\item{numCores}{a integer specifying number of cores of the cluster}

\item{clusterType}{a character string indicating the type of cluster to be created.
If running on personal computer, type should be SOCK. If running on HPC,
type should be MPI}

\item{mainDir}{a character string indicating the path to the directory where
the template and the model link are located, and in which the function will create
the directory structure for the outputs}
}
\value{
an \linkS4class{LPJSetup} object
}
\description{
This function first creates a setup for running the LPJ-GUESS in parallel
 based on the provided input parameters. The function assumes a specific initial
 configuration. A folder (mainDir) containing the model templates for LPJ-GUESS
 (optional) and link to the model executable.
}
\details{
Please note that the requested number of cores should not be larger
than the number of aimed model runs.
}
\examples{
\dontrun{
# 20 different parameters or parameters conbinations to test
  # as a matrix
parameterList <- as.matrix(seq(1,5, len = 20))
colnames(parameterList) <- "common_emax"
  # as a list
dummy <- list (common_emax = NULL)
par <- seq(1,5, len = 20)
parameterList <- vector("list", length(par))
for (i in 1:length(par)) {
  dummy$common_emax <- par[i]
  parameterList[[i]] <- dummy
}


# Create a parallel setup for MPI cluster
# Number of cores must be equal or larger than the parameter combinations
mySetup <- setupLPJParallel(numCores= 20, clusterType = "MPI", mainDir=mainDir)

# Create a parallel setup for SOCK cluster
# Number of cores must be equal or larger than the parameter combinations
mySetup <- setupLPJParallel(numCores= 3, clusterType = "SOCK", mainDir=mainDir)

# Run LPJ-GUESS in parallel
result <- runLPJ(x = mySetup,  settings= settings, parameterList = parameterList )

# After running runLPJ (20 simulations)
str(result,1)

  List of 20

  $ :Formal class 'LPJData' [package "rLPJGUESS"] with 2 slots

  $ :Formal class 'LPJData' [package "rLPJGUESS"] with 2 slots

  [...]



}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig
}
\seealso{
\code{\link{runLPJ}},  \linkS4class{LPJSetup}, \url{https://cran.r-project.org/web/packages/Rmpi/Rmpi.pdf},
 \url{https://cran.r-project.org/web/packages/snow/snow.pdf}
}
\keyword{rLPJGUESS}

