% File LPJ/man/vcf.Rd
\name{vcf}
\alias{vcf}
\title{Fractional forest cover (vegetation continuous field, VCF) or
  MODIS MOD44B.}
\description{
  Half degree data frame of fractional forest cover in percent.
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  \url{https://lpdaac.usgs.gov/products/modis_products_table/mod44b}
}
