% File LPJ/man/wise_v3.Rd
\name{wise_v3}
\alias{wise_v3}
\title{Dataset on global soil depth based on ISRIC WISE version 3.0.}
\description{
Half degree resolution data frame soil depth from ISRIC WISE version 3.0,
converted to NetCDF and further processed in R.
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  \url{http://www.isric.org/data/isric-wise-global-data-set-derived-soil-properties-05-05-degree-grid-ver-30}
  \url{https://github.com/joergsteinkamp/GRASS-GIS_global_soil_properties}
}
