% File LPJ/man/worldmap.Rd
\name{worldmap}
\alias{worldmap}
\title{Dataset on global national borders.}
\description{
  Dataset of world national borders prepared for usage with ggplot2. The
  original dataset is available at \url{http://thematicmapping.org/}.
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
  \url{http://thematicmapping.org/downloads/world_borders.php}
}
