% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTemplate.R
\name{writeTemplate}
\alias{writeTemplate}
\title{A writing template function for LPJ-GUESS}
\usage{
writeTemplate(template1, parameterList, runDir, check = "serial")
}
\arguments{
\item{template1}{a character string providing the general model template,
e.g, global.ins. Provide only the file name, not the path}

\item{parameterList}{a named list containing the parameters to be calibrated}

\item{runDir}{a character string indicating path to the run directory}

\item{check}{a character string indicating how to check the provided parameterList. Default
value is serial. Other possible value is parallel.}
}
\value{
none
}
\description{
This function reads a template, and replaces parameters with the
provides parameters list. If no parameter values are provided, the function will
set them to the default values. The function assumes that a copy of the template
 is already placed in the run directory.
}
\details{
If check is serial, it will return the complete and checked parameterList.
If parallel, it would only check the provided parameters. Please only use serial,
other options are handled internally in the parallelization.
}
\note{
Based an older code of Istem Fer, Uni Potsdam
}
\section{Warning}{
  The provided template can be either one provided by the package or
a self edited template. The function assumes a specific coding for writing the
parameters values. For this reason, we recommend to use the package templates.
If using self edited templates, please take the package templates as a reference.
}

\examples{
\dontrun{
writeTemplate("global.ins", list(run_lamda_max = 0.5, run_emax= 5),
              "/home/lpjRun/runDirectory1")
}
}
\seealso{
\code{\link{getTemplate}}, \code{\link{getParameterList}}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig,
}
\keyword{rLPJGUESS}
