% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getLPJData}
\alias{getLPJData}
\title{A function to process LPJ-GUESS outputs}
\usage{
getLPJData(x, typeList = NULL, runInfo = NULL, processing = FALSE)
}
\arguments{
\item{x}{a character string indicating path to the output files}

\item{typeList}{a character vector with the outputs to be analyzed.
Default value is all}

\item{runInfo}{a named list with the information of the LPJ-GUESS run
The runInfo it will be stored by the function as RData along with
the processed outputs of the model (optional)}

\item{processing}{a boolean indicating whether output files will be turned
into time series (default is FALSE)}
}
\value{
the processed data returned in a S4 Class: LPJData Class
}
\description{
This function reads the ASCII outputs produced by LPJ-GUESS.
It takes a list of output types (typeList) and finds them in the specified path.
The data is stored in a data class object: LPJData. If processing is TRUE,
the data will be stored as zoo time series. Ohterwise, as data frames.
}
\details{
To convert the outputs into zoo time series is only supported when running the
model for one grid cell. For running  LPJ-GUESS for several grid cells, please set
processing to FALSE.
}
\note{
Based on an older code of Joerg Steinkamp
}
\examples{
\dontrun{
LPJout <- getLPJData( typeList = c("aaet",  "cflux","lai", "nflux"),
           "~/path/to/output/files", runInfo = list(parameter1 = 0.5, grid = 1))
}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig
}
\seealso{
\linkS4class{LPJData}, \url{https://cran.r-project.org/web/packages/zoo/zoo.pdf}
}

