% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameterList.R
\name{getParameterList}
\alias{getParameterList}
\title{A get parameter list function}
\usage{
getParameterList(scale, list = TRUE)
}
\arguments{
\item{scale}{a character string indicating whether the parameters are for the global or europe template}

\item{list}{a character boolean to specify the returned format. If TRUE the parameter will be returned
as a list, otherwise as a matrix.}
}
\value{
a list or a matrix with parameter names and their values
}
\description{
This function returns the LPJ-GUESS parameter list required for the writeTemplate function.
It also allows users to see the default values of the templates.
}
\details{
To obtain the parameters used in previous runs, see \code{\link{getRunInfo}}.
}
\examples{
\dontrun{
# Get the default parameters as a list
parameterList <- getParameterList("global", list = TRUE)
parameterList <- getParameterList("europe", list = TRUE)

# Modify any parameter
parameterList$common_emax <- 1
parameterList[["common_emax"]] <- 1


# Get the default parameters as a matrix
parameterList <- getParameterList("global", list = FALSE)
parameterList <- getParameterList("europe", list = FALSE)

# Modify any parameter
parameterList["common_emax",] <- 1

# Call runLPJ with the desired parameter
result <-  runLPJ(x=mainDir, settings=settings, parameterList = parameterList)

}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara
}
\seealso{
\code{\link{runLPJ}}, \code{\link{getRunInfo}}
}

