% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTemplate.R
\name{getTemplate}
\alias{getTemplate}
\title{A get template function}
\usage{
getTemplate(type, outputDir = NULL)
}
\arguments{
\item{type}{a character string indicating the template name: global,
global_cf, global_cru, europe, europe_cf, europe_cru}

\item{outputDir}{a character string indicating path to the output directory  (optional)}
}
\value{
a template object or template.ins file in the specified folder
}
\description{
This function returns the LPJ-GUESS templates.
}
\examples{
\dontrun{
# Get templates as r objects
template <- getTemplate("global")
head(template)
template <- getTemplate("global_cf")
head(template)
template <- getTemplate("global_cru")
head(template)
template <- getTemplate("europe")
head(template)
template <- getTemplate("europe_cf")
head(template)
template <- getTemplate("europe_cru")
head(template)


# Get templates written in a folder
getTemplate("global","/home/LPJTemplates/")
getTemplate("global_cf","/home/LPJTemplates/")
getTemplate("global_cru","/home/LPJTemplates/")
getTemplate("europe","/home/LPJTemplates/")
getTemplate("europe_cf","/home/LPJTemplates/")
getTemplate("europe_cru", "/home/LPJTemplates/")
)
}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig
}
\keyword{rLPJGUESS}

