% File LPJ/man/isleapyear.Rd
\name{is.leapyear}
\alias{is.leapyear}
\title{Is a given year a leap year}
\description{
  Checks whether a given year is a leap year
}
\usage{
is.leapyear(year, always=FALSE)
}
\arguments{
  \item{year}{year to check}
  \item{always}{if TRUE, also relevant years before 1582 will be treated as leap years.}
}
\value{
  boolean
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
