% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLPJData.R
\name{plotLPJData}
\alias{plotLPJData}
\title{A plot function for LPJData objects}
\usage{
plotLPJData(x, typeList = NULL, outDir = NULL, save.plots = FALSE,
  prefix = "")
}
\arguments{
\item{x}{a LPJData object.}

\item{typeList}{a character vector with the outputs to be plotted}

\item{outDir}{a character string indicating the folder where the plots will be
saved, if save.plot set to TRUE}

\item{save.plots}{a boolean indicating whether the plots are saved in the outDir.
Plots will be saved as pdf}

\item{prefix}{a character string specifying the prefix to be added to the plots files.
Only relevant if saving plots is TRUE}
}
\value{
plots for data types included in typeList. The grid cells will be plotted independently
}
\description{
This function reads data from a LPJData object and plots the
 variables against time. If the save.plots is set to TRUE, plots are saved in the output folder.
}
\examples{
\dontrun{
plotLPJData(data, typeList = c("aaet", "lai"),
 outDir = "/runDir/outDir", save.plots = FALSE)
}
}
\seealso{
\url{https://cran.r-project.org/web/packages/zoo/zoo.pdf}
}
\author{
Ramiro Silveyra Gonzalez, Maurizio Bagnara, Florian Hartig
}
\keyword{rLPJGUESS}
