% File LPJ/man/rfs.Rd
\name{rfs}
\alias{rfs}
\title{Rainfall seasonlality, based on CRU (1977-2006) and PGF (1977-2006).}
\description{
Half degree resolution data frame containing rainfall seasonality based
on two climate datasets: CRU and Justin Sheffields
"Princeton Global Forcing" (PGF).
}
\author{
  Joerg Steinkamp <joergsteinkamp@yahoo.de>
}
\references{
Mitchell, T. D. and Jones, P. D.: An improved method of constructing a database of monthly climate observations and associated high-resolution grids, Int. J. Climatol., 25(6), 693-712, doi:10.1002/joc.1181, 2005.
Sheffield, J., Goteti, G. and Wood, E. F.: Development of a 50-Year High-Resolution Global Dataset of Meteorological Forcings for Land Surface Modeling, J. Climate, 19(13), 3088-3111, doi:10.1175/JCLI3790.1, 2006.
}
